+++ swarm/src/lib.rs
@@ -383,13 +383,16 @@ where TBehaviour: NetworkBehaviour<ProtocolsHandler = THandler>,
             return Err(DialError::Banned)
         }
 
-        let mut addrs = me.behaviour.addresses_of_peer(peer_id).into_iter();
-        let peer = me.network.peer(peer_id.clone());
+        let self_listening = &me.listened_addrs;
+        let mut addrs = me.behaviour.addresses_of_peer(peer_id)
+            .into_iter()
+            .filter(|a| !self_listening.contains(a));
 
         let result =
             if let Some(first) = addrs.next() {
                 let handler = me.behaviour.new_handler().into_node_handler_builder();
-                peer.dial(first, addrs, handler)
+                me.network.peer(peer_id.clone())
+                    .dial(first, addrs, handler)
                     .map(|_| ())
                     .map_err(DialError::ConnectionLimit)
             } else {
@@ -497,6 +500,11 @@ where TBehaviour: NetworkBehaviour<ProtocolsHandler = THandler>,
                     let peer = connection.peer_id().clone();
                     let connection = connection.id();
                     this.behaviour.inject_event(peer, connection, event);
+                },
+                Poll::Ready(NetworkEvent::AddressChange { connection, new_endpoint, old_endpoint }) => {
+                    let peer = connection.peer_id();
+                    let connection = connection.id();
+                    this.behaviour.inject_address_change(&peer, &connection, &old_endpoint, &new_endpoint);
                 },
                 Poll::Ready(NetworkEvent::ConnectionEstablished { connection, num_established }) => {
                     let peer_id = connection.peer_id().clone();
@@ -697,11 +705,14 @@ where TBehaviour: NetworkBehaviour<ProtocolsHandler = THandler>,
                             // ongoing dialing attempt, if there is one.
                             log::trace!("Condition for new dialing attempt to {:?} not met: {:?}",
                                 peer_id, condition);
+                            let self_listening = &this.listened_addrs;
                             if let Some(mut peer) = this.network.peer(peer_id.clone()).into_dialing() {
                                 let addrs = this.behaviour.addresses_of_peer(peer.id());
                                 let mut attempt = peer.some_attempt();
-                                for addr in addrs {
-                                    attempt.add_address(addr);
+                                for a in addrs {
+                                    if !self_listening.contains(&a) {
+                                        attempt.add_address(a);
+                                    }
                                 }
                             }
                         }