+++ protocols/ping/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.20.0 [2020-07-01]
+
+- Updated dependencies.
+
+# 0.19.3 [2020-06-22]
+
+- Updated dependencies.
+
 # 0.19.2 [2020-06-18]
 
 - Close substream in inbound upgrade